+++ btcstaking/scripts_utils.go
@@ -11,7 +11,7 @@ import (
 )
 
 // private helper to assemble multisig script
-// if `withVerify` is ture script will end with OP_NUMEQUALVERIFY otherwise with OP_NUMEQUAL
+// if `withVerify` is true script will end with OP_NUMEQUALVERIFY otherwise with OP_NUMEQUAL
 // SCRIPT: <Pk1> OP_CHEKCSIG <Pk2> OP_CHECKSIGADD <Pk3> OP_CHECKSIGADD ... <PkN> OP_CHECKSIGADD <threshold> OP_NUMEQUALVERIFY (or OP_NUMEQUAL)
 func assembleMultiSigScript(
 	pubkeys []*btcec.PublicKey,
@@ -82,7 +82,7 @@ func buildMultiSigScript(
 		return nil, fmt.Errorf("no keys provided")
 	}
 
-	if threshold > uint32(len(keys)) {
+	if int(threshold) > len(keys) {
 		return nil, fmt.Errorf("required number of valid signers is greater than number of provided keys")
 	}