+++ x/btccheckpoint/keeper/msg_server.go
@@ -92,8 +92,8 @@ func (ms msgServer) InsertBTCSpvProof(ctx context.Context, req *types.MsgInsertB
 		return nil, err
 	}
 
-	// At this point, the BTC checkpoint is considered the first valid submission
-	// for the epoch.
+	// At this point, the BTC checkpoint is a valid submission and is
+	// not duplicated (first time seeing the pair of BTC txs)
 	// Thus, we can safely consider this message as refundable
 	ms.k.incentiveKeeper.IndexRefundableMsg(sdkCtx, req)