+++ x/btcstaking/keeper/power_dist_change.go
@@ -170,19 +170,37 @@ func (k Keeper) ProcessAllPowerDistUpdateEvents(
 		switch typedEvent := event.Ev.(type) {
 		case *types.EventPowerDistUpdate_BtcDelStateUpdate:
 			delEvent := typedEvent.BtcDelStateUpdate
+			btcDel, err := k.GetBTCDelegation(ctx, delEvent.StakingTxHash)
+			if err != nil {
+				panic(err) // only programming error
+			}
 			if delEvent.NewState == types.BTCDelegationStatus_ACTIVE {
 				// newly active BTC delegation
-				btcDel, err := k.GetBTCDelegation(ctx, delEvent.StakingTxHash)
-				if err != nil {
-					panic(err) // only programming error
-				}
+
 				// add the BTC delegation to each restaked finality provider
 				for _, fpBTCPK := range btcDel.FpBtcPkList {
 					fpBTCPKHex := fpBTCPK.MarshalHex()
 					activeBTCDels[fpBTCPKHex] = append(activeBTCDels[fpBTCPKHex], btcDel)
 				}
 			} else if delEvent.NewState == types.BTCDelegationStatus_UNBONDED {
-				// add the expired BTC delegation to the map
+				sdkCtx := sdk.UnwrapSDKContext(ctx)
+				// delegation expired and become unbonded emit block event about this
+				// information
+				if btcDel.IsUnbondedEarly() {
+					unbondedEarlyEvent := types.NewDelegationUnbondedEarlyEvent(delEvent.StakingTxHash)
+
+					if err := sdkCtx.EventManager().EmitTypedEvent(unbondedEarlyEvent); err != nil {
+						panic(fmt.Errorf("failed to emit event the new unbonded BTC delegation: %w", err))
+					}
+				} else {
+					expiredEvent := types.NewExpiredDelegationEvent(delEvent.StakingTxHash)
+
+					if err := sdkCtx.EventManager().EmitTypedEvent(expiredEvent); err != nil {
+						panic(fmt.Errorf("failed to emit event for the new expired BTC delegation: %w", err))
+					}
+				}
+
+				// add the unbonded BTC delegation to the map
 				unbondedBTCDels[delEvent.StakingTxHash] = struct{}{}
 			}
 		case *types.EventPowerDistUpdate_SlashedFp: